// enums/Quests.ts
export enum QuestTags {
  ConnectX = 'connect_X',
  ConnectWallet = 'connect_wallet',
  ConnectTelegram = 'connect_Telegram',
  MintSoulbound = 'mint_soulbound',
  SubscribeToNomadz = 'subscribe_to_nomadz',
  Repost = 'repost',
  InviteFriend = 'invite_friend',
  UploadAvatar = 'upload_avatar',
  ApplyToWhitelist = 'apply_to_whitelist',
  ClaimCrossroadsEvent = 'claim_crossroads_event',
  ClaimCountry = 'claim_country',
} 